requirements.txt
PySide6
yara-x
yaraast[full]
capstone
pyinstaller
pycryptodome